refactor: add server route on http
