+++ src/root/root.controller.h
@@ -1,19 +1,15 @@
 #ifndef ROOT_CONTROLLER_H
 #define ROOT_CONTROLLER_H
 
-#include "../server/server.h"
-#include "../shared/router/routes.h"
+#include "../../shared/http/server/server.h"
+#include "../../shared/http/router/routes.h"
 
 typedef struct RootService RootService;
 
-// Route handlers
 extern void handleGetRoot(void* server, void* clientConnection);
 extern void handleGetApi(void* server, void* clientConnection);
 
-// Route getter
 const RouteConfig* getRootRoutes(int* count);
-
-// Service getter
 RootService* getRootService(void);
 
 #endif
