+++ src/root/root.service.c
@@ -1,6 +1,6 @@
 #include "root.service.h"
 
-#include "../shared/formats/json/json.h"
+#include "../../shared/formats/json/json.h"
 
 #include <string.h>
 #include <stdlib.h>
