+++ src/shared/formats/json/json.h
@@ -26,4 +26,4 @@ void jsonEscapeString(const char* input, char* output, size_t outputSize);
  */
 int createJsonObject(char* buffer, size_t bufferSize, int count, ...);
 
-#endif // JSON_UTILS_H
+#endif
