+++ src/shared/http/response/response.c
@@ -5,7 +5,7 @@
 #include <string.h>
 
 // Platform 
-#include "../platform/platform.h"
+#include "../../platform/platform.h"
 
 #define SEND_FUNC send
