+++ src/shared/http/response/response.h
@@ -4,7 +4,7 @@
 #include <stddef.h>
 
 // Platform 
-#include "../platform/platform.h"
+#include "../../platform/platform.h"
 
 /**
  * @brief Sends an HTTP response to the client
