+++ src/shared/http/server/server.h
@@ -6,10 +6,10 @@
 #include <stdint.h>
 
 // Platform 
-#include "../shared/platform/platform.h"
+#include "../../platform/platform.h"
 
 // Include route configuration before Server struct
-#include "../shared/router/route.decorators.h"
+#include "../router/route.decorators.h"
 
 // Constants
 #define MAX_ROUTES 50
@@ -24,27 +24,22 @@ typedef struct Server {
     int routeCount;
 } Server;
 
-// Server lifecycle
 int initializeServer(Server* server, int port);
 void cleanupServer(Server* server);
 void serverListening(Server* server, void (*clientHandler)(Server*, PLATFORM_SOCKET));
 PLATFORM_SOCKET acceptConnection(Server* server);
 
-// Route management
 int addRoute(Server* server, const char* path, RouteHandlerFunc handler);
 int addRouteMethod(Server* server, const char* path, RouteHandlerFunc handler, int method);
 
-// Request handling
 void handleRequest(Server* server, PLATFORM_SOCKET clientConnection, const char* request);
 
-// Response helpers
 void sendResponse(PLATFORM_SOCKET clientConnection, const char* response);
 void sendJsonResponse(PLATFORM_SOCKET clientConnection, const char* json);
 void sendError(PLATFORM_SOCKET clientConnection, int statusCode, const char* message);
 void sendErrorResponse(PLATFORM_SOCKET clientConnection, int statusCode, const char* statusText, const char* message);
 void closeConnection(PLATFORM_SOCKET clientConnection, const char* message);
 
-// Network initialization
 void initSockets(void);
 void cleanupSockets(void);
