+++ src/system/system.module.c
@@ -2,25 +2,22 @@
 #include "system.controller.h"
 #include "system.service.h"
 
-#include "../server/server.h"
-#include "../shared/router/routes.h"
-#include "../shared/http/response.h"
+#include "../../shared/http/server/server.h"
+#include "../../shared/http/router/routes.h"
+#include "../../shared/http/response/response.h"
 
 #include <stdlib.h>
 
-// Service instance
 static SystemService* systemService = NULL;
 
 void systemModuleInit(Server* server) {
     if (!server) return;
     
-    // Initialize service
     systemService = (SystemService*)malloc(sizeof(SystemService));
     if (!systemService) return;
     
     systemServiceInit(systemService);
     
-    // Register all routes from controller
     int count = 0;
     const RouteConfig* routes = getSystemRoutes(&count);
     if (routes && count > 0) {
