+++ src/system/system.module.h
@@ -1,16 +1,13 @@
 #ifndef SYSTEM_MODULE_H
 #define SYSTEM_MODULE_H
 
-#include "../server/server.h"
-#include "../shared/router/routes.h"
+#include "../../shared/http/server/server.h"
+#include "../../shared/http/router/routes.h"
 
-// Initialize the system module
 void systemModuleInit(Server* server);
 
-// Cleanup the system module
 void systemModuleCleanup(Server* server);
 
-// Get the system routes
 const RouteConfig* getSystemRoutes(int* count);
 
 #endif
